Double to Int
Double to int

+++ Food.java
@@ -1,6 +1,6 @@
 public class Food extends Thing{
 
-    public Food(String name, String description, int healthvalue, double x, double y) {
+    public Food(String name, String description, int healthvalue, int x, int y) {
         super(name, description, healthvalue, x, y);
         this.name = "Hamburger";
         this.description = "You obtained a hamburger! 25HP added to your health!";
